desloppify: fix error boundary import cycle and move tooltip to components
Break the mutual import cycle between error-boundary.tsx and
error-boundary-fallback.tsx by moving ErrorBoundaryFallbackProps into the
fallback file where it belongs. error-boundary.tsx now re-exports the type
for backward compatibility.

Move apps/web/src/lib/threat-chart-tooltip.tsx to apps/web/src/components/,
resolving the layer inversion (lib was importing from components and hooks).
Update all import paths accordingly.

Wire Sentry captureException in ErrorBoundary.componentDidCatch to capture
component tree errors — Sentry was already initialized in main.tsx but the
integration was blocked on a stale TODO comment.
