+++ apps/web/src/components/threat-chart-tooltip.tsx
@@ -7,18 +7,18 @@ import { renderToString } from 'react-dom/server'
 import {
   ThreatChartTooltipContent,
   type ThreatChartTooltipContentData,
-} from '../components/threat-chart-tooltip-content'
+} from './threat-chart-tooltip-content'
 import type { ThreatChartThemeColors } from '../hooks/use-threat-chart-theme-colors'
 import type { ThreatSeries } from '../types/app'
-import { resolveSpellSchoolColor } from './spell-school-colors'
-import type { TooltipPointPayload } from './threat-chart-types'
-import { resolveThreatStateStatus } from './threat-chart-visuals'
+import { resolveSpellSchoolColor } from '../lib/spell-school-colors'
+import type { TooltipPointPayload } from '../lib/threat-chart-types'
+import { resolveThreatStateStatus } from '../lib/threat-chart-visuals'
 
 export {
   bossMeleeMarkerColor,
   deathMarkerColor,
   tranquilAirTotemMarkerColor,
-} from './threat-chart-tooltip-colors'
+} from '../lib/threat-chart-tooltip-colors'
 
 function formatResourceTypeLabel(resourceType: number | null): string | null {
   const byResourceTypeCode: Record<number, string> = {
